# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenLUTRA",
    "description": "ROS2 topic recorder for teleoperation robots (ROS2-standard topics)",
    "version": "0.4.3"
  },
  "paths": {
...
    "/api/system/update/apply": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Apply System Update",
        "description": "Apply the pending software update. The application restarts shortly after.",
        "operationId": "applySystemUpdate",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUpdateStatus"
                }
              }
            }
          },
          "409": {
            "description": "Recording in progress, or no pending update"
          },
          "503": {
            "description": "Update agent unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SystemUpdateStatus": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "applying": {
            "type": "boolean",
            "title": "Applying"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "available",
          "applying"
        ],
        "title": "SystemUpdateStatus",
        "description": "Response for GET /api/system/update and POST /api/system/update/apply.\n\n`enabled` is false when no update agent is configured; the UI hides its\nupdate affordances in that case."
      },
...
    }
  }
}